# backend/openapi.yaml
openapi: 3.0.3
info:
  title: BANKERCHANGER API
  version: 0.2.0
  description: REST API for the BANKERCHANGER decentralized boxing prediction market.
servers:
  - url: http://localhost:3001
    description: Local development
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    oracleKey:
      type: apiKey
      in: header
      name: X-Oracle-Key
  schemas:
    Error:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
    ErrorResponse400:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, description: "Bad request. Invalid query parameters, request body, or format." }
        code: { type: string, example: "INVALID_REQUEST" }
      required: [success, message]
    ErrorResponse401:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, description: "Missing, invalid, or expired authentication credentials." }
        code: { type: string, example: "UNAUTHORIZED" }
      required: [success, message]
    ErrorResponse403:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, description: "Authenticated but insufficient permissions for this resource." }
        code: { type: string, example: "FORBIDDEN" }
      required: [success, message]
    ErrorResponse429:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, description: "Rate limit exceeded. Check Retry-After header for reset time." }
        code: { type: string, example: "RATE_LIMIT_EXCEEDED" }
        retryAfter: { type: integer, description: "Seconds until the rate limit window resets." }
      required: [success, message]
    ErrorResponse500:
      type: object
      properties:
        success: { type: boolean, example: false }
        message: { type: string, description: "Internal server error. Please try again later." }
        code: { type: string, example: "INTERNAL_SERVER_ERROR" }
        requestId: { type: string, description: "Unique identifier for this error instance for debugging." }
      required: [success, message]
    UserDTO:
      type: object
      properties:
        id:               { type: string }
        email:            { type: string, format: email }
        emailVerified:    { type: boolean }
        twoFactorEnabled: { type: boolean }
        createdAt:        { type: string, format: date-time }
...
paths:
...
  /api/users/me:
    get:
      summary: Get authenticated user's profile
      operationId: getMyProfile
      tags: [Users]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/UserDTO' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse401' }
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse429' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse500' }
    patch:
      summary: Update authenticated user's profile
      operationId: updateMyProfile
      tags: [Users]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 30
                  pattern: '^[a-zA-Z0-9_]+$'
                avatarUrl: { type: string, format: uri }
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/UserDTO' }
        '400':
          description: Invalid profile data
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse400' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse401' }
        '409':
          description: Username already taken
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse429' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse500' }
...
  /api/users/{id}:
    get:
      summary: Get public user profile
      operationId: getProfile
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Public user profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:    { $ref: '#/components/schemas/UserDTO' }
        '400':
          description: Invalid user ID format
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse400' }
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse429' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse500' }
  /api/users:
    get:
      summary: List all users (admin only)
      operationId: listUsers
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: role
          in: query
          schema:
            type: string
            enum: [BEGINNER, ADVANCED, EXPERT, LEGENDARY]
        - name: status
          in: query
          schema:
            type: string
            enum: [active, suspended]
        - name: search
          in: query
          schema: { type: string }
      responses:
        '200':
          description: Paginated user list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items: { $ref: '#/components/schemas/UserDTO' }
                      total: { type: integer }
                      page:  { type: integer }
                      limit: { type: integer }
        '400':
          description: Invalid query parameters
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse400' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse401' }
        '403':
          description: Admin access required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse403' }
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse429' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse500' }
  /api/users/{id}/suspend:
    patch:
      summary: Suspend a user account (admin only)
      operationId: suspendUser
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: User suspended
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse401' }
        '403':
          description: Admin access required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse403' }
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse429' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse500' }
  /api/users/{id}/role:
    patch:
      summary: Update user role (admin only)
      operationId: updateUserRole
      tags: [Users]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              properties:
                role:
                  type: string
                  enum: [BEGINNER, ADVANCED, EXPERT, LEGENDARY]
      responses:
        '200':
          description: Role updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: object
                    properties:
                      id:   { type: string }
                      tier: { type: string }
        '400':
          description: Invalid role
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse400' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse401' }
        '403':
          description: Admin access required
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse403' }
        '404':
          description: User not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '429':
          description: Rate limit exceeded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse429' }
        '500':
          description: Internal server error
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse500' }

  # ---------------------------------------------------------------------------
  # Platform
  # ---------------------------------------------------------------------------
...